Reworked genome coding
Genomes are no longer read in pairs every time; only instructions that need a
following value cause a read of the following value. The values are also now
parsed into an array of integers rather than parsing the text each time.

The java source is generated and saved in the data directory as well as the
geneticcode.rna