Written RobotLeague
RobotLeague encapsulates all the battles and scoring that concern the bots. I'm
still not particularly happy with the way in which ScoreKeeper works - that
needs fiddling with, but should be well enough hidden that I can do that later
without breaking things.

I've also made the number of evolvebots much less hardcoded; it should be
possible to expand it without too much tinkering if required.