Run for 15k generations against SuperSpinBot
Same setup as before, but run until performance had clearly levelled off. Turns
out to be at about same level as before; the slightly about 50% score is
slightly misleading as this is the best of each generation therefore benefits
from the most favourable random fluctuation - they only tend to score a little
abouve the average of 25%ish.

I suspect that genetic diversity has been destroyed as the behaviour seems to
change very little once it has established the 'drive back and forth spinning
the gun and shooting when it sees an enemy' tactic.