Added selection pressure for genomes shorter than 250 bytes
Added code to reduce the scores of robots with genomes longer than 250
characters (after toString which includes some commas); it's not a true
reflection of the nanobot limit, but it's at least in the right ballpark.

I've impimented it as a separate function as it will almost certianly be heavily
altered until I find the best way for it to work; I might also add a series of
adjustment functions at some point.

Also reduced the starting lengths of genomes as otherwise lots of generations are
run with no selective pressure apart from smaller genomes.